Ingest more storage service properties, fix share encrypted
