+++ CHANGELOG.md
@@ -8,6 +8,18 @@ and this project adheres to
 
 ## [Unreleased]
 
+## 3.13.0 - 2020-04-17
+
+### Added
+
+- Additional properties on `azure_storage_*_service` entities: `kind`,
+  `enableHttpsTrafficOnly`.
+
+### Fixed
+
+- `azure_storage_file_service.encrypted` property was reflecting the `blob`
+  service encryption setting.
+
 ## 3.12.0 - 2020-04-16
 
 ### Added
