+++ src/converters/index.ts
@@ -3,4 +3,3 @@ export * from "./relationships";
 export * from "./resources/compute";
 export * from "./resources/network";
 export * from "./resources/database";
-export * from "./resources/storage";
