+++ src/synchronizers/synchronizeStorageAccounts/converters.ts
@@ -84,8 +84,10 @@ export function createStorageServiceEntity(
         webLink: webLinker.portalResourceUrl(`${data.id}${config.pathSuffix}`),
         region: data.location,
         resourceGroup: resourceGroupName(data.id),
+        kind: data.kind,
         sku: data.sku?.name,
         endpoints: [endpoint],
+        enableHttpsTrafficOnly: data.enableHttpsTrafficOnly,
         category: ["infrastructure"],
       },
       tagProperties: ["environment"],
@@ -147,7 +149,7 @@ export function createStorageFileShareEntity(
         webLink: webLinker.portalResourceUrl(data.id),
         resourceGroup: resourceGroupName(data.id),
         classification: null,
-        encrypted: !!account.encryption?.services?.blob?.enabled,
+        encrypted: !!account.encryption?.services?.file?.enabled,
       },
     },
   });
